Flyttar exception-klassar dit dei høyrer heime
